refactor: findPath lines로 이동
